DateRenderer was getting bloated. Split the code and created PickSleepDate Component for easier readability and loose coupling.
